[Search]: 解决mem chunk出现死锁的问题 - 将数据类型由uint8_t改为uint32_t

+++ src/core/mem_chunk.c
@@ -90,7 +90,7 @@ mem_chunk_t *mem_chunk_init(int num, size_t size)
  ******************************************************************************/
 void *mem_chunk_alloc(mem_chunk_t *chunk)
 {
-    uint8_t i, j;
+    uint32_t i, j;
     uint32_t *bitmap;
 
     spin_lock(&chunk->lock);
